tweaks so that we don't need to swap executables around
I'm including local.settings.json, despite the fact that it's in
.gitignore and that the docs explicitly say you shouldn't check it in.
That's because it might contain secrets.  But it doesn't in this case,
and future modifications to it will be ignored unless explicitly
un-ignored like I'm doing here.  And the executable handling is too
useful not to make this universal.