log directly to Datadog, when configured
This might not be the right answer; it'd be better to just spew to
stdout like always, and have something outside the program ship the logs
to Datadog.  But since we're in a function, there's no agent to do that
for us.  Something about logging to blob storage and having Datadog pick
that up?  Need to investigate.